files/manifest

+++ lib/firefox_zip/parses/manifest.rb
@@ -4,8 +4,10 @@ module FirefoxZip
   module Parses
     class Manifest
       attr_reader :name, :description, :launch_path, :icons, :icon, :type, :permissions
-      
+
       def initialize(file_path)
+        return unless File.exists? file_path
+        
         file = File.open(file_path, 'r')
         data = JSON.parse(file.read)
